Innerclass commit.
